+++ lib/src/crisp_web_sdk.dart
@@ -1,0 +1,2 @@
+export 'crisp_web_sdk_stub.dart'
+    if (dart.library.html) 'crisp_web_sdk_web.dart';